+++ src/main/java/cl/prestabanco/api/models/MorocitiesEntity.java
@@ -18,7 +18,7 @@ public class MorocitiesEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, unique = true)
-    private Long idMorocity;
+    private Integer idMorocity;
     @Column(nullable = false)
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
     private Date dateStartMorocity;
