+++ src/main/java/cl/prestabanco/api/repositories/AddressRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressRepository extends JpaRepository<AddressesEntity, Long> {
+public interface AddressRepository extends JpaRepository<AddressesEntity, Integer> {
     public AddressesEntity findByStreetAddressAndNumberAddressAndCommuneAddressAndRegionAddressAndCountryAddress(
             String streetAddress,
             Integer numberAddress,
