+++ src/main/java/cl/prestabanco/api/repositories/DocumentsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentsRepository extends JpaRepository<DocumentsEntity, Long> {
+public interface DocumentsRepository extends JpaRepository<DocumentsEntity, Integer> {
 }
