+++ src/main/java/cl/prestabanco/api/repositories/RequestsRepository.java
@@ -3,5 +3,5 @@ package cl.prestabanco.api.repositories;
 import cl.prestabanco.api.models.RequestsEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RequestsRepository extends JpaRepository<RequestsEntity, Long> {
+public interface RequestsRepository extends JpaRepository<RequestsEntity, Integer> {
 }
