+++ src/main/java/cl/prestabanco/api/repositories/WithdrawalsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WithdrawalsRepository extends JpaRepository<WithdrawalsEntity, Long> {
+public interface WithdrawalsRepository extends JpaRepository<WithdrawalsEntity, Integer> {
 }
